gptel-ollama: Parse tool results from buffer
* gptel-ollama.el (gptel--parse-buffer):
- Create tool call and result messages from each recorded call

Parse tool messages in the buffer.  This commit covers the
Ollama backend.  The strategy is to pcase on the gptel
property.  Since tool insertions have one distinct gptel text
property, the call and response message are recovered at the same
time.

Introduce support for the `ignore' property in the Ollama
prompt creator.  Users can make use of it to have non-conversation
content intermingled with chat.

NOTE: Ollama does not use tool-call ids, so no ids are stored in
the gptel text property.  This can cause problems if the user
switches from Ollama to a backend that does use ids.