gptel: Prepare reasoning block handling for Deepseek
Depending on the API, there are two ways that reasoning or
chain-of-thought content can appear: as part of the main response
but surrounded by <think>...</think> tags (method 1), or as a
separate JSON field in the response stream (method 2).  Handle
both cases via dispatch on the value of the :reasoning key in
INFO.  :reasoning has five valid values:

- nil before we've checked for <think> blocks or reasoning JSON
  fields.  This applies to both methods.

- 'in when inside a <think> block.  This applies to method 1.

- a string containing the reasoning content (found by the parser
  from a separate JSON field).  This applies to method 2.

- t for the end of the reasoning part of the stream (found by the
  parser from a separate JSON field).  This applies to method 2, and
  only for streaming responses.  This behavior mirrors
  gptel-request's streaming API, which signals that a response
  stream has ended by passing t to the callback.

- 'done if the reasoning content is missing, or done being parsed.
  This applies to both methods.

* gptel.el (gptel--url-get-response): Handle reasoning blocks
  found via method 2.

* gptel-curl.el (gptel-curl--sentinel, gptel-curl--stream-filter,
  gptel-curl--stream-insert-response): Handle reasoning blocks found
  via method 2, and via method 2 + streaming responses.

NOTE: There is no support yet for a backend (like Deepseek) that
uses method 2 to provide reasoning blocks.